--- pkg/subcmd/kubectl/kubectl_suite_test.go
@@ -1,13 +1,0 @@
-package kubectl_test
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
-)
-
-func TestKubectl(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Kubectl Suite")
-}
